Move XcParams,structs to new package
