+++ goavpipe/goavpipe.go
@@ -1,0 +1,4 @@
+// Package goavpipe contains the parts of the avpipe library that may be used by other go packages
+// without relying on the actual libavpipe that does the transcoding. In particular, packages that
+// process the transcode parameters may rely on this to avoid needing a more complex toolchain.
+package goavpipe
